Project no longer maintained
I do not intend to continue maintining linkspam, and have shut down the
Toolforge webservice. The data was last updated on 2021-05, and it needs
some dependency updates and to be moved off the Grid. I don't have the
time to do that at the moment, especially since the logs indicate no
one's actually using it.